Module eventually::aggregate::repository

Module containing the definition of a Repository, to fetch and store Aggregate Roots from a data store.

If you are looking for the Event-sourced implementation of an Aggregate Repository, take a look at EventSourced.

Re-exports

pub use self::any::AnyRepository;

pub use self::any::AnyRepositoryExt;

pub use self::event_sourced::EventSourced;

Modules

any

Contains an super-type that decorates a Repository implementation to return opaque errors using anyhow::Error.

event_sourced

Contains the impementation of the EventSourced Repository trait, necessary to use the Event Sourcing pattern to rehydrate an Aggregate state using an event::Store.

Enums

GetError

Error returned by a call to Repository::get. This type is used to check whether an Aggregate Root has been found or not.

Traits

Repository

A Repository is an object that allows to load and save an [Aggregate Root][Root] from and to a persistent data store.