Trait eventually::aggregate::repository::Repository

pub trait Repository<T>: Send + Syncwhere
    T: Aggregate,{
    type GetError: Send + Sync;
    type SaveError: Send + Sync;

    // Required methods
    fn get<'life0, 'life1, 'async_trait>(
        &'life0 self,
        id: &'life1 T::Id
    ) -> Pin<Box<dyn Future<Output = Result<Root<T>, GetError<Self::GetError>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn save<'life0, 'life1, 'async_trait>(
        &'life0 self,
        root: &'life1 mut Root<T>
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::SaveError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
}

A Repository is an object that allows to load and save an [Aggregate Root][Root] from and to a persistent data store.

Required Associated Types

type GetError: Send + Sync

Error type returned by the concrete implementation of the trait. It is returned in [get] using [GetError::Other].

type SaveError: Send + Sync

Error type returned by the concrete implementation of the trait.

Required Methods

fn get<'life0, 'life1, 'async_trait>( &'life0 self, id: &'life1 T::Id ) -> Pin<Box<dyn Future<Output = Result<Root<T>, GetError<Self::GetError>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Loads an Aggregate Root instance from the data store, referenced by its unique identifier.

fn save<'life0, 'life1, 'async_trait>( &'life0 self, root: &'life1 mut Root<T> ) -> Pin<Box<dyn Future<Output = Result<(), Self::SaveError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Saves a new version of an Aggregate Root instance to the data store.