eventually::aggregate::repository

Trait Saver

pub trait Saver<T>: Send + Sync
where
    T: Aggregate,
{
    // Required method
    fn save<'life0, 'life1, 'async_trait>(
        &'life0 self,
        root: &'life1 mut Root<T>,
    ) -> Pin<Box<dyn Future<Output = Result<(), SaveError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
}

Trait used to implement write access to a data store, which can be used to save the latest state of an aggregate::Root instance.

Required Methods

fn save<'life0, 'life1, 'async_trait>( &'life0 self, root: &'life1 mut Root<T>, ) -> Pin<Box<dyn Future<Output = Result<(), SaveError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Saves a new version of an aggregate::Root instance to the data store.