Struct eventually::aggregate::EventSourcedRepository

pub struct EventSourcedRepository<T, R, S> where
    T: Aggregate,
    R: Root<T>,
    S: Store<StreamId = T::Id, Event = T::Event>,  { /* fields omitted */ }

An Event-sourced implementation of the Repository interface.

It uses an Event Store instance to stream Domain Events for a particular Aggregate, and append uncommitted Domain Events recorded by an Aggregate Root.