Trait eventually::aggregate::Root

pub trait Root<T>: From<Context<T>> + Borrow<Context<T>> + BorrowMut<Context<T>> + Send + Sync where
    T: Aggregate, {
    fn ctx(&self) -> &Context<T> { ... }

    fn ctx_mut(&mut self) -> &mut Context<T> { ... }

    fn aggregate_id<'a>(&'a self) -> &'a T::Id
    where
        T: 'a,
    { ... }
}

An Aggregate Root represents the Domain Entity object used to load and save an Aggregate from and to a Repository, and to perform actions that may result in new Domain Events to change the state of the Aggregate.

An Aggregate Root implementation should only depend on Context, and implement the From<Context<AggregateType>> trait. The Aggregate state and list of Domain Events recorded are handled by the Context object itself.

#[derive(Debug, Clone)]
struct MyAggregateRoot(Context<MyAggregate>);

impl From<Context<MyAggregate>> for MyAggregateRoot {
    fn from(ctx: Context<MyAggregate>) -> Self {
        Self(ctx)
    }
}

// Implement the Aggregate Root interface by providing
// read/write access to the Context object.
impl aggregate::Root<MyAggregate> for MyAggregateRoot {
    fn ctx(&self) -> &Context<MyAggregate> {
        &self.0
    }

    fn ctx_mut(&mut self) -> &mut Context<MyAggregate> {
        &mut self.0
    }
}

For more information on how to record Domain Events using an Aggregate Root, please check Context::record_that method documentation.

Provided methods

fn ctx(&self) -> &Context<T>

Provides read access to an Aggregate Root Context.

fn ctx_mut(&mut self) -> &mut Context<T>

Provides write access to an Aggregate Root Context.

fn aggregate_id<'a>(&'a self) -> &'a T::Id where
T: 'a,

Convenience method to resolve the Aggregate unique identifier from the Aggregate Root instance.