eventually::command

Module test

Module exposing a test Scenario type to write Domain Commands test cases using the given-then-when canvas.

Structs

Scenario

A test scenario that can be used to test a Command Handler using a given-then-when canvas approach.