eventually::event
Module store

Contains implementations of the event::Store trait and connected abstractions, such as the std::collections::HashMap’s based InMemory Event Store implementation.

Structs

InMemory

In-memory implementation of event::Store trait, backed by a thread-safe std::collections::HashMap.

Tracking

Decorator type for an event::Store implementation that tracks the list of recorded Domain Events through it.

Enums

AppendError

All possible error types returned by Appender::append.

Traits

Appender

Interface used to append new Domain Events in an Event Store.

EventStoreExt

Extension trait that can be used to pull in supertypes implemented in this module.

Store

An Event Store, used to store Domain Events in Event Streams – a stream of Domain Events – and retrieve them.

Streamer

Interface used to stream Persisted Domain Events from an Event Store to an application.