eventually::event::store
Trait Appender

pub trait Appender<StreamId, Event>: Send + Sync
where
    StreamId: Send + Sync,
    Event: Message + Send + Sync,
{
    // Required method
    fn append<'life0, 'async_trait>(
        &'life0 self,
        id: StreamId,
        version_check: Check,
        events: Vec<Envelope<Event>>,
    ) -> Pin<Box<dyn Future<Output = Result<Version, AppendError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Interface used to append new Domain Events in an Event Store.

Required Methods

fn append<'life0, 'async_trait>( &'life0 self, id: StreamId, version_check: Check, events: Vec<Envelope<Event>>, ) -> Pin<Box<dyn Future<Output = Result<Version, AppendError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Appens new Domain Events to the specified Event Stream.

The result of this operation is the new Version of the Event Stream with the specified Domain Events added to it.