Trait eventually::event::Appender

pub trait Appender<StreamId, Event>: Send + Syncwhere
    StreamId: Send + Sync,
    Event: Message + Send + Sync,{
    type Error: Into<Option<ConflictError>> + Send + Sync;

    fn append<'life0, 'async_trait>(
        &'life0 self,
        id: StreamId,
        version_check: StreamVersionExpected,
        events: Vec<Envelope<Event>>
    ) -> Pin<Box<dyn Future<Output = Result<Version, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

Interface used to append new Domain Events in an Event Store.

Required Associated Types

type Error: Into<Option<ConflictError>> + Send + Sync

The error type returned by the Store during an [append] call. It could be a [version::ConflictError], which is why the bound to Into<Option<ConflictError>> is required.

Required Methods

fn append<'life0, 'async_trait>(
&'life0 self,
id: StreamId,
version_check: StreamVersionExpected,
events: Vec<Envelope<Event>>
) -> Pin<Box<dyn Future<Output = Result<Version, Self::Error>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Appens new Domain Events to the specified Event Stream.

The result of this operation is the new Version of the Event Stream with the specified Domain Events added to it.