Trait eventually::event::Store

pub trait Store<StreamId, Event>: Streamer<StreamId, Event> + Appender<StreamId, Event> + Send + Sync where
    StreamId: Send + Sync,
    Event: Message + Send + Sync, { }

An [Event] Store, used to store Domain Events in Event Streams – a stream of Domain Events – and retrieve them.

Each Event Stream is represented by a unique Stream identifier.