Trait eventually::event::Streamer

pub trait Streamer<StreamId, Event>: Send + Sync where
    StreamId: Send + Sync,
    Event: Message + Send + Sync, {
    type Error: Send + Sync;

    fn stream(
        &self,
        id: &StreamId,
        select: VersionSelect
    ) -> Stream<'_, StreamId, Event, Self::Error>;
}

Interface used to stream Persisted Domain Events from an Event Store to an application.

Required Associated Types

type Error: Send + Sync

The error type returned by the Store during a [stream] call.

Required Methods

fn stream(
&self,
id: &StreamId,
select: VersionSelect
) -> Stream<'_, StreamId, Event, Self::Error>

Opens an Event Stream, effectively streaming all Domain Events of an Event Stream back in the application.