Struct eventually::optional::AsAggregate

pub struct AsAggregate<T>(_);

Adapter for Aggregate types to the foundational eventually::Aggregate trait.

Examples

use eventually_util::optional::Aggregate as OptionalAggregate;

enum SomeEvent {
    Happened
}

#[derive(Debug, PartialEq)]
struct SomeState {
    // Some nice fields
}

struct SomeAggregate;
impl OptionalAggregate for SomeAggregate {
    type State = SomeState;
    type Event = SomeEvent;
    type Error = std::convert::Infallible;

    fn apply_first(event: Self::Event) -> Result<Self::State, Self::Error> {
        // Return an empty state, here you should create the initial
        // state based on the event received.
        Ok(SomeState {})
    }

    fn apply_next(state: Self::State, _event: Self::Event) -> Result<Self::State, Self::Error> {
        // Return the same state, here you should update the state
        // based on the event received.
        Ok(state)
    }
}

use eventually_core::aggregate::Aggregate;
use eventually_util::optional::AsAggregate;

// To adapt SomeAggregate to `eventually::Aggregate`:
let result = AsAggregate::<SomeAggregate>::apply(
    None,                   // This state will result in calling `SomeAggregate::apply_first`
    SomeEvent::Happened,
);

// An `Option`-wrapped `SomeState` instance is returned.
assert_eq!(result, Ok(Some(SomeState {})));