eventually::query

Type Alias Envelope

pub type Envelope<T> = Envelope<T>;

A Message carrying the Domain Query itself as payload and other relevant information as metadata.

Aliased Type

struct Envelope<T> {
    pub message: T,
    pub metadata: HashMap<String, String>,
}

Fields
message: T

The message payload.

metadata: HashMap<String, String>

Optional metadata to provide additional context to the message.