eventually
Module serde

This module provides traits and implementations for serialization and deserialization, allowing you to convert Rust data structures to and from different formats like JSON, Protobuf, etc.

Structs

Convert

Implements the Serde trait to translate between two different types, and using the specified Serde for serialization and deserialization using the new Out type.

Json

Implements the Serializer and Deserializer traits, which use the serde crate to serialize and deserialize a message into JSON.

ProtoJson

Implementation of Serde traits that uses ProtoJson as wire protocol.

Protobuf

Implements the Serde trait which serializes and deserializes the message using Protobuf format through the prost::Message trait.

Traits

Deserializer

A deserializer interface that can be used to deserialize a byte array into an instance of a specific Rust data type from a specific wire format.

Serde

Serializer and Deserializer that can be used to serialize into and deserialize from a given type into a specific wire format, such as JSON, Protobuf, etc.

Serializer

A serializer interface that can be used to serialize a Rust data type into a specific wire format as a byte array.