eventually::serde
Trait Deserializer

pub trait Deserializer<T>: Send + Sync {
    // Required method
    fn deserialize(&self, data: &[u8]) -> Result<T>;
}

A deserializer interface that can be used to deserialize a byte array into an instance of a specific Rust data type from a specific wire format.

Required Methods

fn deserialize(&self, data: &[u8]) -> Result<T>

Deserializes the given value from a message encoded in the wire format supported by this implementation.

Errors

An error (anyhow::Error) is returned in case the deserialization could not succeed as expected.