Trait eventually::serde::Serde

pub trait Serde<T>: Send + Sync {
    type Error: Send + Sync;

    // Required methods
    fn serialize(&self, value: T) -> Vec<u8>;
    fn deserialize(&self, data: Vec<u8>) -> Result<T, Self::Error>;
}

A Serde can be used to serialize into and deserialize from a given type into a wire format, such as JSON or Protobuf.

Required Associated Types

type Error: Send + Sync

The error returned by the Serde::deserialize method.

Required Methods

fn serialize(&self, value: T) -> Vec<u8>

Serializes the given value into the wire format supported by this Serde.

fn deserialize(&self, data: Vec<u8>) -> Result<T, Self::Error>

Deserializes the given value from a message encoded in the wire format supported by this Serde.