Struct eventually::Repository

pub struct Repository<T, Store> where
    T: Aggregate + Clone + 'static,
    Store: EventStore<SourceId = <T as Aggregate>::Id, Event = <T as Aggregate>::Event>,  { /* fields omitted */ }

Implementation of the Repository pattern for storing, retrieving and deleting Aggregates.

A Repository instruments an EventStore to:

Insert Events in the EventStore for an Aggregate, using the AggregateRoot,

Get all the Events in the EventStore and rebuild the State of an Aggregate, into a new AggregateRoot instance,

Remove all the Events for an Aggregate in the EventStore.

Implementations

impl<T, Store> Repository<T, Store> where
T: Aggregate + Clone,
Store: EventStore<SourceId = <T as Aggregate>::Id, Event = <T as Aggregate>::Event>,

pub fn new(
builder: AggregateRootBuilder<T>,
store: Store
) -> Repository<T, Store>

Creates a new Repository instance, using the Aggregate and EventStore provided.

impl<T, Store> Repository<T, Store> where
T: Aggregate + Clone,
Store: EventStore<SourceId = <T as Aggregate>::Id, Event = <T as Aggregate>::Event>,
<T as Aggregate>::Id: Debug,
<T as Aggregate>::Id: Clone,
<T as Aggregate>::Event: Clone,
<T as Aggregate>::Error: Error,
<T as Aggregate>::Error: 'static,
<Store as EventStore>::Error: Error,
<Store as EventStore>::Error: 'static,

pub async fn get(
&'_ self,
id: <T as Aggregate>::Id
) -> Result<AggregateRoot<T>, Error<<T as Aggregate>::Error, <Store as EventStore>::Error>>

Returns the Aggregate from the Repository with the specified id, if any.

In case the Aggregate with the specified id exists, returns a new AggregateRoot instance with its latest State.

Otherwise, None is returned.

pub async fn add(
&'_ mut self,
root: AggregateRoot<T>
) -> Result<AggregateRoot<T>, Error<<T as Aggregate>::Error, <Store as EventStore>::Error>>

Adds a new State of the Aggregate into the Repository, through an AggregateRoot instance.

Returns Error::NoEvents if there are no Events to commit in the AggregateRoot.

pub async fn remove(
&'_ mut self,
id: <T as Aggregate>::Id
) -> Result<(), Error<<T as Aggregate>::Error, <Store as EventStore>::Error>>

Removes the specified Aggregate from the Repository, using the provided AggregateId.