Trait eventually::AggregateExt

pub trait AggregateExt: Aggregate {
    fn fold<I>(
        state: Self::State, 
        events: I
    ) -> Result<Self::State, Self::Error>
    where
        I: Iterator<Item = Self::Event>,
    { ... }

    fn async_fold<'a, S>(
        state: Self::State, 
        events: S
    ) -> Pin<Box<dyn Future<Output = Result<Self::State, Self::Error>> + 'a + Send>>
    where
        S: Stream<Item = Self::Event> + Send + 'a,
        Self::State: Send,
        Self::State: 'a,
        Self::Event: Send,
        Self::Event: 'a,
        Self::Error: Send,
        Self::Error: 'a,
    { ... }
}

Extension trait for Aggregate containing combinator functions.

Provided methods

fn fold<I>(state: Self::State, events: I) -> Result<Self::State, Self::Error> where
I: Iterator<Item = Self::Event>,

Applies a synchronous stream of Events to the current State, returning the updated state or an error, if any such happened.

fn async_fold<'a, S>(
state: Self::State,
events: S
) -> Pin<Box<dyn Future<Output = Result<Self::State, Self::Error>> + 'a + Send>> where
S: Stream<Item = Self::Event> + Send + 'a,
Self::State: Send,
Self::State: 'a,
Self::Event: Send,
Self::Event: 'a,
Self::Error: Send,
Self::Error: 'a,

Applies an asynchronous Stream of Events to the current State, returning the updated state or an error, if any such happened.

Loading content...