Trait eventually::Store

pub trait Store {
    type SourceId: PartialEq<Self::SourceId>;
    type Offset: PartialOrd<Self::Offset>;
    type Event;
    type Error;
    fn stream(
        &self, 
        source_id: Self::SourceId, 
        from: Self::Offset
    ) -> Pin<Box<dyn Stream<Item = Result<Self::Event, Self::Error>> + Send>>;

    fn append(
        &mut self, 
        source_id: Self::SourceId, 
        events: Vec<Self::Event>
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send>>;
}

Represents an Event Store, an append-only, ordered list of Events for a certain "source" (i.e. Aggregate).

Associated Types

type SourceId: PartialEq<Self::SourceId>

Type of the Source id, usually an Aggregate id.

type Offset: PartialOrd<Self::Offset>

Type of the memory offset supported by the Store.

An offset is needed to get a slice of the events in the Store.

Check out Store::stream for more information.

type Event

Type of the events supported by the Store.

Usually, they match the same type as Aggregate::Event.

type Error

Possible errors returned by the Store.

Usually, this should be an enum containing all the possible reasons why the Store could fail due to external failures.

Loading content...
Required methods

fn stream(
&self,
source_id: Self::SourceId,
from: Self::Offset
) -> Pin<Box<dyn Stream<Item = Result<Self::Event, Self::Error>> + Send>>

Allows to stream many Events from the Store back to the application, by specifying the SourceId and the desired Offset.

An asynchronous Stream is returned, yielding every Event from the specified Offset in the same order as they were committed.

To stream back all the events in the Store for a certain Source, Offset should implement Default, and use such value (in case of numeric offsets, this is most likely 0).

In case of an Aggregate, it is possible to recreate its State by supplying the returned Stream into Aggregate::async_fold.

fn append(
&mut self,
source_id: Self::SourceId,
events: Vec<Self::Event>
) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send>>

Appends a list of new events to the Store.

An Error is returned if the append operation fails.

Loading content...