Crate eventually

Modules

aggregate

inmemory

optional

repository

store

versioned

Structs

AggregateRoot
An AggregateRoot represents an handler to the Aggregate it's managing, such as:

AggregateRootBuilder
Builder type for new AggregateRoot instances.

Repository
Implementation of the Repository pattern for storing, retrieving and deleting Aggregates.

Traits

Aggregate
An Aggregate manages a domain entity State, acting as a transaction boundary.

AggregateExt
Extension trait with some handy methods to use with Aggregates.

EventStore
An Event Store is an append-only, ordered list of Events for a certain "source" -- e.g. an Aggregate.

Type Definitions

AggregateId
A short extractor type for the Aggregate Id.