Crate eventually

eventually is a crate that helps you apply different patterns to your Rust application domain code, such as: Event Sourcing, Aggregate Root, Outbox Pattern, and so on.

Modules

aggregate

Module containing support for the Aggregate pattern.

command

Module containing support for Domain [Command]s.

event

Module event contains types and abstractions helpful for working with Domain Events.

message

This module contains the definition of a Message type, which can be used to describe some sort of domain value such as a Domain Event, a Domain Command, and so on.

query

Module query contains types and helpful abstractions to model Domain Queries and implement Domain Query Handlers.

serde

This module provides traits and implementations for serialization and deserialization, allowing you to convert Rust data structures to and from different formats like JSON, Protobuf, etc.

tracing

Module containing some extension traits to support code instrumentation using the tracing crate.

version

Contains the types necessary for Optimistic Locking through versioning.