Module eventually::aggregate

Structs

AggregateRoot
An AggregateRoot represents an handler to the Aggregate it's managing, such as:

Versioned
Newtype extension for Aggregate types to add support for Optimistic Concurrency Control using versioning for Events and Aggregate State.

Traits

Aggregate
An Aggregate manages a domain entity State, acting as a transaction boundary.

AggregateExt
Extension trait with some handy methods to use with Aggregates.

Identifiable
A trait for data structures that can be identified by an id.

Optional
An Option-flavoured, Aggregate-compatible trait to model Aggregates having an optional State.

Type Definitions

AggregateId
A short extractor type for the Aggregate id, found in the Aggregate State.