Module eventually::aggregate

Modules

referential
Contains Aggregate variant where aggregate root and State are implemented by the same object.

Traits

Aggregate
An Aggregate is an entity which State is composed of one or more value-objects, entities or nested aggregates.

AggregateExt
Extension trait for Aggregate containing combinator functions.

Identifiable

ReferentialAggregate
An Aggregate representation where the State is the same as the aggregate root.

Type Definitions

ErrorOf
Alias for the Error type of an Aggregate.

EventOf
Alias for the Event type of an Aggregate.

StateOf
Alias for the State type of an Aggregate.