Trait eventually::aggregate::repository::Getter

pub trait Getter<T>: Send + Syncwhere
    T: Aggregate,{
    type Error: Send + Sync;

    // Required method
    fn get<'life0, 'life1, 'async_trait>(
        &'life0 self,
        id: &'life1 T::Id
    ) -> Pin<Box<dyn Future<Output = Result<Root<T>, GetError<Self::Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
}

Getter is a Repository trait used to get an aggregate::Root instance from a data store.

Required Associated Types

type Error: Send + Sync

Error type returned by the concrete implementation of the trait. It is returned in [get] using [GetError::Other].

Required Methods

fn get<'life0, 'life1, 'async_trait>( &'life0 self, id: &'life1 T::Id ) -> Pin<Box<dyn Future<Output = Result<Root<T>, GetError<Self::Error>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Loads an Aggregate Root instance from the data store, referenced by its unique identifier.