Trait eventually::aggregate::repository::Getter

pub trait Getter<T>: Send + Sync
where
    T: Aggregate,
{
    // Required method
    fn get<'life0, 'life1, 'async_trait>(
        &'life0 self,
        id: &'life1 T::Id
    ) -> Pin<Box<dyn Future<Output = Result<Root<T>, GetError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
}

Trait used to implement read access to a data store from which to load an aggregate::Root instance, given its id.

Required Methods

fn get<'life0, 'life1, 'async_trait>( &'life0 self, id: &'life1 T::Id ) -> Pin<Box<dyn Future<Output = Result<Root<T>, GetError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Loads an aggregate::Root instance from the data store, referenced by its unique identifier.