eventually

Module event

Module event contains types and abstractions helpful for working with Domain Events.

Re-exports

pub use crate::event::store::Store;

Modules

store

Contains implementations of the event::Store trait and connected abstractions, such as the std::collections::HashMap’s based InMemory Event Store implementation.

Structs

Persisted

An [Event] that has been persisted to the Event Store.

Enums

VersionSelect

Specifies the slice of the Event Stream to select when calling [Store::stream].

Type Aliases

Envelope

An Event is a Message carring the information about a Domain Event, an occurrence in the system lifetime that is relevant for the Domain that is being implemented.

Stream

Stream is a stream of Persisted Domain Events.