Module eventually::event

Module event contains types and abstractions helpful for working with Domain Events.

Structs

Persisted

An Event that has been persisted to the Event Store.

Enums

StreamVersionExpected

Specifies an expectation on the Event Stream version targeted when calling Store::append.

VersionSelect

Specifies the slice of the Event Stream to select when calling Store::stream.

Traits

Store

An Event Store, used to store Domain Events in Event Streams – a stream of Domain Events – and retrieve them.

Type Definitions

Event

An Event is a Message carring the information about a Domain Event, an occurrence in the system lifetime that is relevant for the Domain that is being implemented.

Stream

Stream is a stream of Persisted Domain Events.