Module eventually::event

Module event contains types and abstractions helpful for working with Domain Events.

Modules

store

Contains implementations of the event::Store trait and connected abstractions, such as the std::collections::HashMap’s based InMemory Event Store implementation.

Structs

Persisted

An [Event] that has been persisted to the Event Store.

Enums

StreamVersionExpected

Specifies an expectation on the Event Stream version targeted when calling [Store::append].

VersionSelect

Specifies the slice of the Event Stream to select when calling [Store::stream].

Traits

Appender

Interface used to append new Domain Events in an Event Store.

Store

An [Event] Store, used to store Domain Events in Event Streams – a stream of Domain Events – and retrieve them.

Streamer

Interface used to stream Persisted Domain Events from an Event Store to an application.

Type Aliases

Envelope

An Event is a [Message] carring the information about a Domain Event, an occurrence in the system lifetime that is relevant for the Domain that is being implemented.

Stream

Stream is a stream of Persisted Domain Events.