Trait eventually::event::Store

pub trait Store: Send + Sync {
    type StreamId: Send + Sync;
    type Event: Message + Send + Sync;
    type StreamError: Send + Sync;
    type AppendError: Into<Option<ConflictError>> + Send + Sync;

    fn stream(
        &self,
        id: &Self::StreamId,
        select: VersionSelect
    ) -> Stream<'_, Self::StreamId, Self::Event, Self::StreamError>;
    fn append<'life0, 'async_trait>(
        &'life0 self,
        id: Self::StreamId,
        version_check: StreamVersionExpected,
        events: Vec<Envelope<Self::Event>>
    ) -> Pin<Box<dyn Future<Output = Result<Version, Self::AppendError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

An [Event] Store, used to store Domain Events in Event Streams – a stream of Domain Events – and retrieve them.

Each Event Stream is represented by a unique Store::StreamId.

Required Associated Types

type StreamId: Send + Sync

The type used to uniquely identify each Event Stream recorded by the Event Store.

type Event: Message + Send + Sync

The type containing all Domain Events recorded by the Event Store. Typically, this parameter should be an enum.

type StreamError: Send + Sync

The error type returned by the Store during a [stream] call.

type AppendError: Into<Option<ConflictError>> + Send + Sync

The error type returned by the Store during an [append] call. It could be a [version::ConflictError], which is why the bound to Into<Option<ConflictError>> is required.

Required Methods

fn stream(
&self,
id: &Self::StreamId,
select: VersionSelect
) -> Stream<'_, Self::StreamId, Self::Event, Self::StreamError>

Opens an Event Stream, effectively streaming all Domain Events of an Event Stream back in the application.

fn append<'life0, 'async_trait>(
&'life0 self,
id: Self::StreamId,
version_check: StreamVersionExpected,
events: Vec<Envelope<Self::Event>>
) -> Pin<Box<dyn Future<Output = Result<Version, Self::AppendError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Appens new Domain Events to the specified Event Stream.

The result of this operation is the new Version of the Event Stream with the specified Domain Events added to it.