Module eventually::repository

Module containing Repository implementation to retrieve, save and delete Aggregates.

Repository and Aggregates

As described in the Interacting with Aggregates using AggregateRoot section in the aggregate module-level documentation, in order to interact with an Aggregate instance you need to use an AggregateRoot.

To get an AggregateRoot, you can use a Repository instance.

The Repository allows to retrieve, save and remove specific Aggregate instances, by using an underlying EventStore implementation that handles the Aggregate’s events.

A Repository will always return an AggregateRoot instance on read, whether or not events are present in the EventStore.

Use the Repository to implement your bounded-context application logic, for example in HTTP or RPC handlers.

Structs

Repository

Implementation of the Repository pattern for storing, retrieving and deleting Aggregates.

Enums

Error

Error type returned by the Repository.

Type Definitions

Result

Result type returned by the Repository.