Trait eventually::Subscription

pub trait Subscription {
    type SourceId: Eq;
    type Event;
    type Error;
    fn resume(
        &self
    ) -> Pin<Box<dyn Future<Output = Result<Pin<Box<dyn Stream<Item = Result<Persisted<Self::SourceId, Self::Event>, Self::Error>> + Send, Global>>, Self::Error>> + Send, Global>>;

    fn checkpoint(
        &self, 
        version: u32
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send, Global>>;
}

A Subscription to an EventStream which can be “checkpointed”: keeps a record of the latest message processed by itself using checkpoint, and can resume working from such message by using the resume.

Associated Types

type SourceId: Eq

Type of the Source id, typically an AggregateId.

type Event

Event type stored in the EventStore, typically an Aggregate::Event.

type Error

Possible errors returned when receiving events from the notification channel.

Required methods

fn resume(
&self
) -> Pin<Box<dyn Future<Output = Result<Pin<Box<dyn Stream<Item = Result<Persisted<Self::SourceId, Self::Event>, Self::Error>> + Send, Global>>, Self::Error>> + Send, Global>>

Resumes the current state of a Subscription by returning the EventStream, starting from the last event processed by the Subscription.

fn checkpoint(
&self,
version: u32
) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send, Global>>

Saves the provided version (or sequence number) as the latest version processed.